backend/requirements.txt
openai==0.27.0
python-dotenv==1.0.0
Pillow==9.5.0
transformers==4.30.0
Flask==3.0.2
Werkzeug==3.0.1
Flask-Cors==4.0.1
opencv-python==4.9.0.80
numpy<2.0.0
gunicorn==20.1.0